pyproject.toml
[project]
name = "flow-fund"
version = "0.1.0"
description = ""
authors = [
    {name = "Achyuth",email = "achyuthrayal2@gmail.com"}
]
requires-python = ">=3.11,<3.15"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn[standard] (>=0.44.0,<0.45.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "openai (>=2.32.0,<3.0.0)",
    "faiss-cpu (>=1.13.2,<2.0.0)",
    "numpy (>=2.4.4,<3.0.0)",
    "langgraph (>=1.1.8,<2.0.0)",
]


[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.3"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
